reorganise component directories
