+++ src/components/GameSetup.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { observer } from "mobx-react";
 import { useMainStore } from "../hooks/useMainStore";
 import { Game } from "./game/Game";
-import { StartOverlay } from "./overlays/StartOverlay";
+import { StartOverlay } from "./start-overlay/StartOverlay";
 import { GameStage } from "../stores/gameStore";
 
 export const GameSetup = observer(() => {
