+++ src/components/game/betting-controls/BettingControls.tsx
@@ -3,11 +3,11 @@ import { observer } from "mobx-react";
 import { BettingChips } from "./betting-chips/BettingChips";
 import { BettingFields } from "./betting-fields/BettingFields";
 import { BettingButtons } from "./betting-buttons/BettingButtons";
-import { GameStage } from "../../stores/gameStore";
-import { useMainStore } from "../../hooks/useMainStore";
-import { useAudio } from "../../hooks/useAudio";
-import { playAudio } from "../../utils";
-import startSound from "../../sounds/place-bets-please.mp3";
+import { GameStage } from "../../../stores/gameStore";
+import { useMainStore } from "../../../hooks/useMainStore";
+import { useAudio } from "../../../hooks/useAudio";
+import { playAudio } from "../../../utils";
+import startSound from "../../../sounds/place-bets-please.mp3";
 
 export const BettingControls = observer(() => {
   const { game, sound } = useMainStore();
