+++ src/components/game/betting-controls/betting-buttons/BettingButtons.module.scss
@@ -1,4 +1,4 @@
-@import '../../../mixins';
+@import '../../../../mixins';
 
 .controls {
   padding: 0 1rem 1rem 1rem;
