+++ src/components/game/betting-controls/betting-chips/BettingChip.module.scss
@@ -1,4 +1,4 @@
-@import '../../../mixins';
+@import '../../../../mixins';
 
 .chip {
   transform: perspective(10cm) rotateX(20deg);
