+++ src/components/game/betting-controls/betting-chips/BettingChips.module.scss
@@ -1,4 +1,4 @@
-@import '../../../mixins';
+@import '../../../../mixins';
 
 .chips {
   display: flex;
