+++ src/components/game/betting-controls/betting-chips/BettingChips.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { observer } from "mobx-react";
 import { BettingChip, BettingChipsValues } from "./BettingChip";
-import { GameStage } from "../../../stores/gameStore";
-import { useMainStore } from "../../../hooks/useMainStore";
+import { GameStage } from "../../../../stores/gameStore";
+import { useMainStore } from "../../../../hooks/useMainStore";
 import styles from "./BettingChips.module.scss";
 
 const bettingChipValues: BettingChipsValues[] = [1, 5, 25, 50, 100];
