+++ src/components/game/bottom-information/BottomInformation.module.scss
@@ -1,4 +1,4 @@
-@import "../../mixins";
+@import "../../../mixins";
 
 .bottom {
   grid-row: 3 / span 2;
