+++ src/components/game/cards/Cards.module.scss
@@ -1,4 +1,4 @@
-@import "../../mixins";
+@import "../../../mixins";
 
 .cards {
   display: flex;
