+++ src/components/game/participant-field/ParticipantField.module.scss
@@ -1,4 +1,4 @@
-@import "../../mixins";
+@import "../../../mixins";
 
 .field {
   grid-row: 3 / 3;
