+++ src/components/game/participant-field/ParticipantField.tsx
@@ -2,11 +2,11 @@ import React from "react";
 import cn from "classnames";
 import { Cards } from "../cards/Cards";
 import styles from "./ParticipantField.module.scss";
-import { Card } from "../../types";
+import { Card } from "../../../types";
 
 type ParticipantFieldProps = {
   prefix: string;
-  participant: string;
+  participant: "player" | "banker";
   points: number;
   cards: Card[];
   revertDirection: boolean;
