+++ src/components/game/timer/Timer.module.scss
