+++ src/components/game/timer/Timer.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { observer } from "mobx-react";
-import { useMainStore } from "../../hooks/useMainStore";
-import { formatTime } from "../../utils";
+import { useMainStore } from "../../../hooks/useMainStore";
+import { formatTime } from "../../../utils";
 import styles from "./Timer.module.scss";
 
 export const Timer = observer(() => {
