+++ src/components/start-overlay/StartOverlay.module.scss
