+++ src/components/start-overlay/instructions/Instructions.module.scss
@@ -1,4 +1,4 @@
-@import "../../mixins";
+@import "../../../mixins";
 
 .instructions {
   margin: 0 auto;
