+++ src/components/start-overlay/instructions/Instructions.tsx
